This commit adds support for Waterloo (GRT) and Vancouver (Translink) where all their vehicles are retrieved and are inserted into the mongo collection, whose name is the timestamp generated in the mongo_helper.py file.
Currently, the program must be invoked manually each time,
and the endpoints for Waterloo and Vancouver are hardcoded.

The agency names are also hardcoded as to disambiguate the
agency to whom a vehicle belongs to. This could be eventually
solved by having different collections for different agencies,
or by keeping the current solution.